build: split extended appchain verification into optional tiers
