+++ .github/workflows/build.yml
@@ -47,6 +47,7 @@ jobs:
         with:
           java-version: '25'
           distribution: 'temurin'
+          cache: 'gradle'
       - name: Grant execute permission for gradlew
         run: chmod +x gradlew
       - name: Build with Gradle
@@ -73,10 +74,73 @@ jobs:
           appchain/appchain-composite-contracts/scripts/verify-cddl.sh
       - name: Verify ADR-013 demo release contracts
         run: bash app/appchain-effects-demo/tests/release-contracts.sh
-      - name: Smoke final JVM plugin catalog and directory loading
-        run: ./gradlew :app:packagedJvmPluginCatalogSmoke -PskipSigning=true --stacktrace
       - name: Verify and catalog-launch standalone plugin scaffold
         run: ./gradlew -p scaffolds/plugin-template clean build -PyanoSource="$PWD" -PskipSigning=true --stacktrace
+
+  appchain-extended:
+    if: >-
+      github.event_name != 'schedule' &&
+      (github.event_name != 'workflow_dispatch' || inputs.scope == 'all')
+    runs-on: ubuntu-latest
+    steps:
+      - name: Checkout
+        uses: actions/checkout@34e114876b0b11c390a56381ad16ebd13914f8d5
+        with:
+          submodules: 'true'
+      - name: Set up JDK 25
+        uses: actions/setup-java@c1e323688fd81a25caa38c78aa6df2d33d3e20d9
+        with:
+          java-version: '25'
+          distribution: 'temurin'
+          cache: 'gradle'
+      - name: Grant execute permission for gradlew
+        run: chmod +x gradlew
+      - name: Run extended app-chain suites
+        run: ./gradlew appChainExtendedTest -PskipSigning=true --stacktrace
+
+  appchain-crypto:
+    if: >-
+      github.event_name != 'schedule' &&
+      (github.event_name != 'workflow_dispatch' || inputs.scope == 'all')
+    runs-on: ubuntu-latest
+    timeout-minutes: 30
+    steps:
+      - name: Checkout
+        uses: actions/checkout@34e114876b0b11c390a56381ad16ebd13914f8d5
+        with:
+          submodules: 'true'
+      - name: Set up JDK 25
+        uses: actions/setup-java@c1e323688fd81a25caa38c78aa6df2d33d3e20d9
+        with:
+          java-version: '25'
+          distribution: 'temurin'
+          cache: 'gradle'
+      - name: Grant execute permission for gradlew
+        run: chmod +x gradlew
+      - name: Run full app-chain cryptographic verification
+        run: ./gradlew appChainCryptoTest -PskipSigning=true --stacktrace
+
+  distribution-check:
+    if: >-
+      github.event_name != 'schedule' &&
+      (github.event_name != 'workflow_dispatch' || inputs.scope == 'all')
+    runs-on: ubuntu-latest
+    timeout-minutes: 30
+    steps:
+      - name: Checkout
+        uses: actions/checkout@34e114876b0b11c390a56381ad16ebd13914f8d5
+        with:
+          submodules: 'true'
+      - name: Set up JDK 25
+        uses: actions/setup-java@c1e323688fd81a25caa38c78aa6df2d33d3e20d9
+        with:
+          java-version: '25'
+          distribution: 'temurin'
+          cache: 'gradle'
+      - name: Grant execute permission for gradlew
+        run: chmod +x gradlew
+      - name: Verify packaged JVM and showcase distributions
+        run: ./gradlew distributionCheck -PskipSigning=true --stacktrace
 
   native-plugin-conformance:
     if: >-
@@ -185,6 +249,9 @@ jobs:
   milestone-1-release-acceptance:
     needs:
       - commit-build
+      - appchain-extended
+      - appchain-crypto
+      - distribution-check
       - native-plugin-conformance
       - connector-fault-matrix
       - effect-failover-e2e
@@ -198,12 +265,18 @@ jobs:
       - name: Require every Milestone 1 release gate
         env:
           COMMIT_BUILD_RESULT: ${{ needs.commit-build.result }}
+          APPCHAIN_EXTENDED_RESULT: ${{ needs.appchain-extended.result }}
+          APPCHAIN_CRYPTO_RESULT: ${{ needs.appchain-crypto.result }}
+          DISTRIBUTION_CHECK_RESULT: ${{ needs.distribution-check.result }}
           NATIVE_PLUGIN_RESULT: ${{ needs.native-plugin-conformance.result }}
           CONNECTOR_FAULT_RESULT: ${{ needs.connector-fault-matrix.result }}
           EFFECT_FAILOVER_RESULT: ${{ needs.effect-failover-e2e.result }}
         run: |
           set -euo pipefail
           test "$COMMIT_BUILD_RESULT" = success
+          test "$APPCHAIN_EXTENDED_RESULT" = success
+          test "$APPCHAIN_CRYPTO_RESULT" = success
+          test "$DISTRIBUTION_CHECK_RESULT" = success
           test "$NATIVE_PLUGIN_RESULT" = success
           test "$CONNECTOR_FAULT_RESULT" = success
           test "$EFFECT_FAILOVER_RESULT" = success
