+++ .github/workflows/release.yml
@@ -15,6 +15,7 @@ jobs:
         with:
           java-version: '25'
           distribution: 'temurin'
+          cache: 'gradle'
       - name: Grant execute permission for gradlew
         run: chmod +x gradlew
       - name: Read project version
@@ -31,7 +32,7 @@ jobs:
           fi
           echo "version=${VERSION}" >> "$GITHUB_OUTPUT"
       - name: Build with Gradle
-        run: ./gradlew clean build
+        run: ./gradlew clean fullBuild
       - name: Decode GPG Key
         run: |
           mkdir -p ~/.gradle/
