+++ .github/workflows/snapshot_manual.yml
@@ -16,10 +16,11 @@ jobs:
         with:
           java-version: '25'
           distribution: 'temurin'
+          cache: 'gradle'
       - name: Grant execute permission for gradlew
         run: chmod +x gradlew
       - name: Build with Gradle
-        run: ./gradlew clean build -PskipSigning=true --stacktrace
+        run: ./gradlew clean fullBuild -PskipSigning=true --stacktrace
       - name: Publish package to SNAPSHOT repository
         run: ./gradlew publishToSonatype -PskipSigning=true --warn --stacktrace
         env:
