+++ appchain/appchain-devtools/src/main/resources/appchain-dx/v1alpha1/appchain-release-acceptance-index.json
@@ -34,14 +34,14 @@
     "profile:eutxo-key-payments": ["appchain/extensions/eutxo/appchain-eutxo-ledger/src/test/java/com/bloxbean/cardano/yano/appchain/eutxo/ledger/EutxoStateMachineTest.java#sameBlockChainingWorksAndFirstConflictingSpendWins"],
     "funding:eutxo-genesis": ["appchain/extensions/eutxo/appchain-eutxo-ledger/src/test/java/com/bloxbean/cardano/yano/appchain/eutxo/ledger/EutxoStateMachineTest.java#sameBlockChainingWorksAndFirstConflictingSpendWins"],
     "bridge:cardano-federated": ["appchain/extensions/eutxo/appchain-eutxo-ledger/src/test/java/com/bloxbean/cardano/yano/appchain/eutxo/ledger/EutxoStateMachineTest.java#stableVaultObservationCreditsExactlyOnceAndUpdatesReserveAtomically", "appchain/extensions/eutxo/appchain-eutxo-ledger/src/test/java/com/bloxbean/cardano/yano/appchain/eutxo/ledger/EutxoStateMachineTest.java#signedWithdrawalBecomesIrrevocableAndExactL1ConfirmationReconcilesReserve", "appchain/extensions/eutxo/appchain-eutxo-bridge-cardano/src/test/java/com/bloxbean/cardano/yano/appchain/eutxo/bridge/cardano/WithdrawalCoordinatorTest.java#crashAfterSubmitRetriesTheSameSignedTransactionWithoutResigning", "appchain/extensions/eutxo/appchain-eutxo-bridge-cardano/src/test/java/com/bloxbean/cardano/yano/appchain/eutxo/bridge/cardano/ProofWithdrawalTransactionBuilderTest.java#permissionlessRelayReturnsExactReusableSignedBytes", "appchain/extensions/eutxo/appchain-eutxo-bridge-onchain/src/test/java/com/bloxbean/cardano/yano/appchain/eutxo/bridge/onchain/ProofVaultValidatorConformanceTest.java#realMpfClaimSettlesPermissionlesslyWithinBudget", "appchain/extensions/eutxo/appchain-eutxo-bridge-onchain/src/test/java/com/bloxbean/cardano/yano/appchain/eutxo/bridge/onchain/RootAndNullifierValidatorConformanceTest.java#nullifierAcceptsOnlyTheExactNextClaimSequence"],
-    "settlement:zeroj-validity": ["appchain/extensions/eutxo-zk/appchain-eutxo-zk-testkit/src/test/java/com/bloxbean/cardano/yano/appchain/eutxo/zk/testkit/EutxoValidityIntegrationTest.java#selectedOptionalProviderMaintainsAtomicSecondRoot", "appchain/extensions/eutxo-zk/appchain-eutxo-zk-onchain/src/test/java/com/bloxbean/cardano/yano/appchain/eutxo/zk/onchain/EutxoValidityRootValidatorTest.java#rootAdvanceRequiresOneExactProofBoundDataPublication", "appchain/extensions/eutxo-zk/appchain-eutxo-zk-zeroj/src/test/java/com/bloxbean/cardano/yano/appchain/eutxo/zk/zeroj/EutxoZ6CircuitHardeningTest.java#deterministicFuzzCorpusSatisfiesCircuitAndMutationsNeverAlias"],
+    "settlement:zeroj-validity": ["appchain/extensions/eutxo-zk/appchain-eutxo-zk-testkit/src/test/java/com/bloxbean/cardano/yano/appchain/eutxo/zk/testkit/EutxoValidityIntegrationTest.java#selectedOptionalProviderMaintainsAtomicSecondRoot", "appchain/extensions/eutxo-zk/appchain-eutxo-zk-onchain/src/cryptoTest/java/com/bloxbean/cardano/yano/appchain/eutxo/zk/onchain/EutxoValidityRootValidatorTest.java#rootAdvanceRequiresOneExactProofBoundDataPublication", "appchain/extensions/eutxo-zk/appchain-eutxo-zk-zeroj/src/cryptoTest/java/com/bloxbean/cardano/yano/appchain/eutxo/zk/zeroj/EutxoZ6CircuitFuzzTest.java#deterministicFuzzCorpusSatisfiesCircuitAndMutationsNeverAlias"],
     "rollup:zeroj-cardano": ["appchain/extensions/eutxo-zk/Z6_PRODUCTION_HARDENING.md#external-gates"],
     "indexer:eutxo-lifecycle": ["appchain/extensions/eutxo/appchain-eutxo-indexer-jdbc/src/test/java/com/bloxbean/cardano/yano/appchain/eutxo/indexer/jdbc/SqliteEutxoIndexStoreTest.java#withdrawalVersionsRollbackWithoutLosingTheDepositHistory", "appchain/extensions/eutxo/appchain-eutxo-indexer-core/src/test/java/com/bloxbean/cardano/yano/appchain/eutxo/indexer/EutxoIndexCoordinatorTest.java#slowStoreCoalescesNotificationBurstsWithoutBlockingFinality"],
-    "sequencer:fixed": ["runtime/src/test/java/com/bloxbean/cardano/yano/runtime/appchain/AppChainSequencingIntegrationTest.java#sequencer_finalizesBlocks_identicalRoots_andSurvivesRestart"],
-    "sequencer:rotating": ["runtime/src/test/java/com/bloxbean/cardano/yano/runtime/appchain/AppChainRotationIntegrationTest.java#rotation_survivesProposerCrash"],
-    "membership:static": ["runtime/src/test/java/com/bloxbean/cardano/yano/runtime/appchain/AppChainTwoNodeIntegrationTest.java#twoNodes_exchangeAuthenticatedMessages_bothDirections"],
-    "membership:governed": ["runtime/src/test/java/com/bloxbean/cardano/yano/runtime/appchain/GovernedMembershipIntegrationTest.java#thresholdApproval_activates_andLateJoinerDerivesHistory"],
-    "l1:slot-feed": ["runtime/src/test/java/com/bloxbean/cardano/yano/runtime/appchain/AppChainRotationIntegrationTest.java#feedSlot"],
+    "sequencer:fixed": ["runtime/src/integrationTest/java/com/bloxbean/cardano/yano/runtime/appchain/AppChainSequencingIntegrationTest.java#sequencer_finalizesBlocks_identicalRoots_andSurvivesRestart"],
+    "sequencer:rotating": ["runtime/src/integrationTest/java/com/bloxbean/cardano/yano/runtime/appchain/AppChainRotationIntegrationTest.java#rotation_survivesProposerCrash"],
+    "membership:static": ["runtime/src/test/java/com/bloxbean/cardano/yano/runtime/appchain/AppChainTwoNodeSmokeTest.java#twoNodes_exchangeAuthenticatedMessages_bothDirections"],
+    "membership:governed": ["runtime/src/integrationTest/java/com/bloxbean/cardano/yano/runtime/appchain/GovernedMembershipIntegrationTest.java#thresholdApproval_activates_andLateJoinerDerivesHistory"],
+    "l1:slot-feed": ["runtime/src/integrationTest/java/com/bloxbean/cardano/yano/runtime/appchain/AppChainRotationIntegrationTest.java#feedSlot"],
     "anchor:metadata": ["runtime/src/test/java/com/bloxbean/cardano/yano/runtime/appchain/AnchorServiceTest.java#observedConfirmationRetriesOnTickWithoutReplayOrResubmit"],
     "anchor:script": ["runtime/src/test/java/com/bloxbean/cardano/yano/runtime/appchain/ScriptAnchorServiceTest.java#First real anchor"],
     "observer:address-deposit": ["runtime/src/test/java/com/bloxbean/cardano/yano/runtime/appchain/L1ObservationServiceTest.java#observers_computeDepositAndMetadataClaims"],
@@ -122,7 +122,7 @@
       "gitOpsTargets": ["helm", "kustomize"],
       "outcomeLevel": "OPERATOR_OWNED_REFERENCE",
       "acceptanceScenario": "recipe-eutxo-zeroj-validity",
-      "evidence": ["appchain/extensions/eutxo-zk/appchain-eutxo-zk-testkit/src/test/java/com/bloxbean/cardano/yano/appchain/eutxo/zk/testkit/EutxoValidityIntegrationTest.java#selectedOptionalProviderMaintainsAtomicSecondRoot", "appchain/extensions/eutxo-zk/appchain-eutxo-zk-onchain/src/test/java/com/bloxbean/cardano/yano/appchain/eutxo/zk/onchain/EutxoValidityRootValidatorTest.java#rootAdvanceRequiresOneExactProofBoundDataPublication", "appchain/extensions/eutxo-zk/Z6_PRODUCTION_HARDENING.md#external-gates"]
+      "evidence": ["appchain/extensions/eutxo-zk/appchain-eutxo-zk-testkit/src/test/java/com/bloxbean/cardano/yano/appchain/eutxo/zk/testkit/EutxoValidityIntegrationTest.java#selectedOptionalProviderMaintainsAtomicSecondRoot", "appchain/extensions/eutxo-zk/appchain-eutxo-zk-onchain/src/cryptoTest/java/com/bloxbean/cardano/yano/appchain/eutxo/zk/onchain/EutxoValidityRootValidatorTest.java#rootAdvanceRequiresOneExactProofBoundDataPublication", "appchain/extensions/eutxo-zk/Z6_PRODUCTION_HARDENING.md#external-gates"]
     },
     {
       "id": "eutxo-zeroj-preview", "maturity": "experimental",
