+++ appchain/appchain-devtools/src/test/resources/appchain-dx/v1alpha1/metadata.sha256
@@ -7,5 +7,5 @@ appchain-gitops-lock.schema.json=2fe0c5ef68c42ddad6df46ecee59834eb7fb9a119353394
 appchain-lock.schema.json=90db95529a50eef1deff98fad66a6732458535f87b840103e3421083ad0e6ce5
 appchain-metadata-trust.schema.json=b0fa6e1336f278c1910812d059316df226fbb40b92b2a6e7e010348a2dd2801b
 appchain-recipe-catalog.json=0e9516894224d6af84e069d6b0769f111fda6740db8f4d5eba7dbc0708077e0e
-appchain-release-acceptance-index.json=f27f5c282cfd468593da789178957079d0c243725fb1dccedba4d563a5f2e070
+appchain-release-acceptance-index.json=21b7f828d64ba5fccc9fd4de18c7a4548c2e11b555987ba989870dd39422e9de
 appchain-release-capability-index.json=21ad2054fafc8dab53fe7ab45f84422b5c2ca2c961a63ccc6a39cb8a352430cd
