+++ appchain/examples/appchain-showcase/src/distributionTest/scripts/distribution-contract.sh
