+++ appchain/extensions/eutxo-zk/appchain-eutxo-zk-onchain/build.gradle
@@ -29,7 +29,20 @@ java {
     targetCompatibility = 25
 }
 
-tasks.named('test') {
+tasks.withType(Test).configureEach {
     systemProperty 'zeroj.allowInsecureTrustedSetup', 'true'
+}
+
+tasks.named('test') {
+    maxHeapSize = '2g'
+}
+
+tasks.named('cryptoTest') {
     maxHeapSize = '4g'
+}
+
+sourceSets.cryptoTest.compileClasspath += sourceSets.test.output
+sourceSets.cryptoTest.runtimeClasspath += sourceSets.test.output
+tasks.named('compileCryptoTestJava') {
+    dependsOn tasks.named('testClasses')
 }
