+++ gradle.properties
@@ -5,3 +5,4 @@ quarkusPluginVersion = 3.33.2.1
 quarkusPlatformGroupId = io.quarkus.platform
 quarkusPlatformArtifactId = quarkus-bom
 quarkusPlatformVersion = 3.33.2.1
+org.gradle.caching = true
